feat(math): 2d math library (#2)
## Summary

2D math library (milestone 1, PR 2 of 5) — `include/silk/math.h`, all
functions `static inline`, header-only:

- **Scalar utilities** — `sl_min/max/clamp/abs`, epsilon comparison
`sl_feq`
- **`sl_vec2`** — add/sub/scale/neg, dot, scalar cross (CCW sign
convention), length/length_sq, zero-safe normalize, CCW perpendicular,
lerp, distance
- **`sl_mat2`** — row-major storage, column-vector convention: identity,
CCW rotation, multiply, transform, transpose
- **Tests** — 15 cases covering scalar edge behavior, vector identities
(orthogonality, zero-safe normalize, cross sign), and matrix properties
(90° rotation, composition == direct 180°, R^T·R == I for pure rotation)

## Motivation

Milestone 1, PR 2 per [ROADMAP.md](../ROADMAP.md): the math foundation
every later PR builds on — simulation state (PR 3) stores positions and
orientations with it, integration (PR 4) transforms them, the sandbox
(PR 5) draws them.

## Changes

- `float` throughout — aligns with the future SoA/GPU data model
- Zero-length normalize returns the zero vector instead of NaN; the
threshold compares **squared** length against `SL_EPSILON` (physics-safe
default, documented as intentionally not scale-invariant)
- Header-only `static inline` keeps hot paths optimizer-friendly without
LTO; functions can move to TUs later if profiling demands
- No convenience alias for applying a matrix (`sl_mat2_mul_vec2(m, v)`
is the one way) — minimal surface area
- Test assertions follow the comparison law from AGENTS.md: exact
equality only where IEEE-754 guarantees it (construction, integer-valued
arithmetic), `SL_EXPECT_NEAR` tolerances everywhere floats flow through
sqrt/trig; rotation cases use a looser `1e-5` epsilon because `SL_PI /
2` rounds
- Suite wired per the four-file checklist, including `add_test(NAME math
COMMAND sl_tests)`
- `silk` links `libm` PUBLIC on UNIX: public headers call
`sqrtf/sinf/cosf`, glibc keeps those in a separate library, and
consumers must get it transitively (macOS folds it into libSystem;
Windows carries it in the CRT)

## Testing

```sh
cmake --preset debug && cmake --build --preset debug && ctest --preset debug   # core + math pass
ctest --preset sanitize                                                        # ASan + UBSan clean
```

`[math]` suite: 15/15 ok alongside `[core]`; header verified
warning-clean under gcc/clang with the full strict flag set.

## Notes

- Conventions: CCW-positive angles; perpendicular is `(-y, x)`; matrices
apply as `M · v`
- Considered `SL_ASSERT` argument checks on math primitives and rejected
them — pure functions over value domains stay assert-free so hot-path
codegen stays identical across builds
- Epsilon comparison stays caller-scaled (`sl_feq` takes eps);
`SL_EPSILON` is only a default constant, not a universal tolerance

+++ tests/silk_test.h
@@ -70,7 +70,7 @@ int sl_run_suite(const char *suite, const sl_test_case *cases, int count)
         double sl_d_ = sl_a_ - sl_b_;                                         \
         if (sl_d_ < 0.0)                                                      \
             sl_d_ = -sl_d_;                                                   \
-        if (!(sl_d_ <= (eps))) {                                              \
+        if (!(sl_d_ <= (double)(eps))) {                                      \
             printf("\n    FAIL %s:%d: |%s - %s| = %g > %g\n",                 \
                    __FILE__, __LINE__, #a, #b, sl_d_, (double)(eps));         \
             sl_test_failures++;                                               \
